requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.1.0
certifi==2022.9.24
colorama==0.4.6
exceptiongroup==1.0.4
iniconfig==1.1.1
packaging==21.3
pip==22.3.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.2.0
pytest-xdist==3.0.2
pytest-rerunfailures==10.3
selenium==4.6.1
setuptools==65.6.3
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
requests~=2.28.2
aiogram==2.25.1
Faker==18.6.0
pydantic==1.10.7
jsonschema==4.17.3